refactor(compiler): move real-world codegen tests into __tests__
The `real-world/` folder name was unclear. These cross-cutting tests compile
fixtures and assert generated output (not tied to one source file), so they live
in a conventional `__tests__/` folder alongside the codegen targets. The
StyledRecipe exemplar moves in too as styled-recipe.test.ts.
