+++ core/compiler/src/codegen/targets/__tests__/composition.test.ts
