+++ core/compiler/src/codegen/targets/__tests__/events.test.ts
