+++ core/compiler/src/codegen/targets/__tests__/lifecycle-effects.test.ts
