+++ core/compiler/src/codegen/targets/__tests__/props.test.ts
