+++ core/compiler/src/codegen/targets/__tests__/refs.test.ts
