+++ core/compiler/src/codegen/targets/__tests__/scoped-slots.test.ts
