+++ core/compiler/src/codegen/targets/__tests__/slots.test.ts
