+++ core/compiler/src/codegen/targets/__tests__/transitions.test.ts
